backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pandas==2.0.3
numpy==1.24.3
psycopg2-binary==2.9.7
scikit-learn==1.3.0
imbalanced-learn==0.11.0
joblib==1.3.2
statsmodels==0.14.0
matplotlib==3.7.2
seaborn==0.12.2
openpyxl==3.1.2
xlrd==2.0.1
